pyproject.toml
[project]
name = "skai-fastmcp-cloudrun"
version = "0.1.0"
description = "FastMCP server for Australian Hansard RAG with ChatGPT Developer Mode support"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.5",
    "langchain>=1.0.0,<2.0.0",
    "langchain-text-splitters>=1.0.0",
    "langchain-google-vertexai>=2.0.0",
    # Temporary: keep wrapper until new backend is fully wired, then remove
    # New vector backend dependencies (016-langchain-postgres-upgrade)
    "langchain-postgres>=0.0.12",
    "sqlalchemy>=2.0.0",
    "cloud-sql-python-connector[pg8000]>=1.4.0",
    "pg8000>=1.30.0",  # Required for Cloud SQL Connector pg8000 driver
    "psycopg[binary]>=3.1,<3.3",  # Pinned to minor version range for stability
    "pgvector>=0.3.0",
    "langgraph",
    "pydantic>=2.0",
    "uvicorn>=0.30.0",
    "pyyaml>=6.0",
    "PyJWT[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",  # Used for Cloud Run metadata service
    "google-cloud-secret-manager>=2.16.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.0.0",
    "fastapi>=0.111.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "pytest-httpx>=0.22.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "manual: Manual tests that require user interaction",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
